// comonl-comonlwcl/src/app/modules/comonlapi/model/datiAzienda.ts
/**
 * ComonlWeb
 * API per ComonlWeb
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ApiError } from './apiError';
import { Atecofin } from './atecofin';
import { Ccnl } from './ccnl';
import { Comune } from './comune';
import { LegaleRappr } from './legaleRappr';
import { NaturaGiuridica } from './naturaGiuridica';
import { Sede } from './sede';
import { StatiEsteri } from './statiEsteri';


export interface DatiAzienda { 
    apiWarnings?: Array<ApiError>;
    id?: number;
    idUserInsert?: string;
    idUserUltMod?: string;
    dtInsert?: Date;
    dtUltMod?: Date;
    capReferente?: string;
    cfAzienda?: string;
    cfCapogruppo?: string;
    cfReferente?: string;
    cognomeReferente?: string;
    denominazioneDatoreLavoro?: string;
    emailReferente?: string;
    faxReferente?: string;
    flgCapogruppoEstero?: string;
    flgProspettoDaCapogruppo?: string;
    indirizzoReferente?: string;
    nomeReferente?: string;
    telefonoReferente?: string;
    atecofin?: Atecofin;
    ccnl?: Ccnl;
    comune?: Comune;
    statiEsteri?: StatiEsteri;
    sedeLegale?: Sede;
    elencoSedi?: Array<Sede>;
    idAziendaSilp?: string;
    cessata?: boolean;
    accentrata?: boolean;
    presenteSuAAEP?: boolean;
    naturaGiuridica?: NaturaGiuridica;
    flagArtigiana?: string;
    codLavoroTemp?: string;
    flagTipoAzienda?: string;
    descrTipoAzienda?: string;
    flagMaster?: string;
    altreInformazioni?: string;
    partitaIva?: string;
    flgPubblicaAmm?: string;
    ilLegaleRapprDellaSedeLegale?: LegaleRappr;
}