// comonl-comonlwcl/src/app/modules/comonlapi/model/datore.ts
/**
 * ComonlWeb
 * API per ComonlWeb
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ApiError } from './apiError';
import { Atecofin } from './atecofin';
import { LegaleRappr } from './legaleRappr';
import { NaturaGiuridica } from './naturaGiuridica';
import { SedeLavoro } from './sedeLavoro';


export interface Datore { 
    apiWarnings?: Array<ApiError>;
    id?: number;
    naturaGiuridica?: NaturaGiuridica;
    codiceFiscale?: string;
    dsCognome?: string;
    dsDenominazioneDatore?: string;
    dsNome?: string;
    dsVariazioneRagSociale?: string;
    dtFineContrattoSom?: Date;
    dtInizioContrattoSom?: Date;
    flgAzArtigiana?: string;
    flgDenomSilpVariata?: string;
    flgNoAaep?: string;
    flgPubAmm?: string;
    flgUtilEstera?: string;
    idAziendaSilp?: number;
    matricolaInps?: string;
    numAgenziaSommin?: string;
    numContrSomm?: number;
    numeroIscrizioneAlbo?: string;
    partitaIva?: string;
    patInail?: string;
    atecofin?: Atecofin;
    legaleRappr?: LegaleRappr;
    sedeLegale?: SedeLavoro;
    sedeOperativa?: SedeLavoro;
}