// comonl-comonlwcl/src/app/modules/comunicazione/modules/nav-main-comunicazione/models/nav-comunicazione-params.ts
/*
* SPDX-FileCopyrightText: Copyright 2019 - 2022 | CSI Piemonte
* SPDX-License-Identifier: EUPL-1.2
*/
export interface NavComunicazioneParams {
    idComunicazione?: number;
    formModalParams?: any;
    mode?: string;
    flgArrivoComeNuova?: boolean;
    quandroIndex?: number;
    configurazioneQuadri?: string;
    flgNuovaCO: boolean;
    cfLavoratorePerVardatore?: string;
}